Timeline fixes (#78)
* feat: Add color stops to transition to the end of the timeline

This commit adds a pair of color stops to the DensityLine component in order to transition smoothly to the end of the timeline when there is a gap between the end of the last period and the end of the timeline. This improves the visual representation of the timeline.

* feat: Filter out zeroish-length segments in TimelineView

This commit filters out zeroish-length segments in the TimelineView component. By checking the difference between the end and start values of each segment, we ensure that only segments with a non-zero length are included. This improves the accuracy of the timeline representation.