feat: Include playlist ID in episode list API response
This commit modifies the `get_list` handler in the `episode` module to include the `playlist_id` field in the API response. The `results` vector now contains a tuple of `(Episode, Option<chrono::NaiveDateTime>, Option<String>)`, where the third element represents the `playlist_id` associated with each episode. This change allows users to retrieve the playlist ID along with other episode details when making requests to the episode list API.