populate other metadata for created youtube videos (#81)
* feat: Add recording date and playlist options to YoutubeUploadRequest

This commit adds the `recording_date`, `playlist_id`, and `playlist_position` fields to the `YoutubeUploadRequest` struct. These fields allow users to specify the recording date of the video and add it to a specific playlist on YouTube.

* feat: Add playlist_id field to series create and update handlers

This commit adds the `playlist_id` field to the `create` and `update` handlers in the `series` module. The `playlist_id` field is now included in the `CreateSeriesRequest` and `UpdateSeriesRequest` structs, allowing users to specify a playlist ID when creating or updating a series. This feature enables users to associate a series with a specific playlist in the application.

* feat: Include playlist ID in episode list API response

This commit modifies the `get_list` handler in the `episode` module to include the `playlist_id` field in the API response. The `results` vector now contains a tuple of `(Episode, Option<chrono::NaiveDateTime>, Option<String>)`, where the third element represents the `playlist_id` associated with each episode. This change allows users to retrieve the playlist ID along with other episode details when making requests to the episode list API.

* feat: Add recording date and playlist options to YoutubeUploadRequest in frontend`

* feat: Add thumbnail_url and playlist_id fields to series create and edit forms

* feat: Update episode and series structs to include ISO 8601 formatted timestamps

This commit updates the `Episode` and `Series` structs in the `structs.rs` file to include ISO 8601 formatted timestamps for the `created_at`, `updated_at`, and `stream_date` fields. The timestamps are formatted using the `%Y-%m-%dT%H:%M:%S%.3fZ` pattern. This change ensures consistent and standardized timestamp representation in the API responses and improves compatibility with other systems that rely on ISO 8601 formatted timestamps.

* feat: Update YouTube video recording date to ISO 8601 format

This commit updates the `upload_video_handler` function in the `main.rs` file to format the `recordingDate` field of the YouTube video upload request as an ISO 8601 timestamp. The timestamp is formatted using the `%Y-%m-%dT%H:%M:%S%.3fZ` pattern. This change ensures that the recording date is consistently represented in the ISO 8601 format, improving compatibility with other systems that rely on this standard.

* feat: Update `upload_video_handler` to handle YouTube playlists

This commit modifies the `upload_video_handler` function in the `main.rs` file to improve the handling of YouTube video uploads. It includes changes to the `upload` function, adding error handling and response handling logic. Additionally, it introduces the `response` variable to capture the result of the upload operation. This update ensures that video uploads are handled correctly and provides better error reporting in case of failures.

* feat: Update Dockerfile.rust to use dynamic Debian version

This commit updates the Dockerfile.rust to use a dynamic Debian version by introducing the `DEBIAN_VERSION` argument. This allows for greater flexibility in choosing the Debian version for the runtime stage. By default, the latest Debian version is used. This change ensures that the Docker image is built with the most up-to-date Debian version, improving compatibility and security.

* feat: Update iso8601 dependency to include chrono feature

This commit updates the `iso8601` dependency in the `Cargo.toml` file to include the `chrono` feature. This change ensures compatibility with the `chrono` library, allowing for the use of ISO 8601 formatted timestamps in the codebase.

* feat: Update `upload_video_handler` to include recording details in YouTube video upload

This commit modifies the `upload_video_handler` function in the `main.rs` file to include the `recordingDetails` field in the YouTube video upload request. The `recordingDetails` field is populated with the `recordingDate` value, formatted as an ISO 8601 timestamp. This change ensures that the recording date is consistently represented in the ISO 8601 format, improving compatibility with other systems that rely on this standard.

* refactor: extract structs from youtube_upload_api/main.rs