Chain tasks (#110)
* Add TaskTemplate/next_task

* chore: Update npm dependencies to latest versions

* chore: Update RUST_LOG level to debug in docker-compose.yaml

* Fix some frontend lint errors/warnings

* chore: Update StreamSimpleView struct to include series_id field

The StreamSimpleView struct in the `handlers/stream/structs.rs` file has been updated to include a new field called `series_id`. This field is an optional string that represents the ID of the series associated with the stream. This change allows for better organization and categorization of streams.

* refactor: Refactor audio extraction in detect_segment function

The audio extraction process in the `detect_segment` function of the `transcription.rs` file has been refactored. Use of unwrap was removed, and some parsing error were changed to return 0 and trace a warning

* Add support for a `next_task` recursive task chain in the backend

* refactor: Update Task component to include new features and improve UI

The Task component in the List.tsx file has been updated to include new features and improve the user interface. Changes include:
- Change UI with a new secondary action button to mark tasks as viewed
- Added a badge to indicate if there are more tasks to start after the current one finishes
- Updated the background color of selected tasks to improve visibility
- Improved the frontend types of tasks