use chained tasks to add videos to playlist after upload to youtube (#159)
* use chained tasks to add videos to playlist after upload to youtube
Fixes #137

* chore: Update Rust channel to 1.80

* feat: Refactor stream_ingestion.rs to handle ffprobe errors

Refactor the `get_entries` function in `stream_ingestion.rs` to handle errors that occur during the ffprobe process. Previously, if an error occurred, the function would continue processing the next file without logging the error. This could lead to silent failures and make debugging difficult.

Now, when an error occurs during the ffprobe process, the function logs the error using the `tracing` crate and continues processing the next file. This ensures that errors are properly logged and developers can easily identify and debug any issues.

This change improves the reliability and maintainability of the code by providing better error handling and logging.

* feat: Update summary field in upload_video_handler

Update the summary field in the upload_video_handler function to be an array instead of an object. This change ensures consistency with the rest of the codebase and improves data structure readability.

* chore: Update Create.tsx to include stream_id and series_id inputs

* feat: Refactor work loop to handle task requeueing and failure

Refactor the work loop in the `main.rs` file to handle task requeueing and failure scenarios. This change introduces a new `WorkLoopResult` enum to represent the possible outcomes of the work loop: `Complete`, `Requeued`, and `Failed`.

The work loop now breaks out of the loop and returns the appropriate `WorkLoopResult` value based on the outcome of each iteration. If the task is requeued, the next task is created and added to the queue. If the task fails, the status is updated accordingly.

This refactoring improves the clarity and maintainability of the code by separating the different outcomes of the work loop and handling them appropriately.

* Refactor handling of previous_task_id in TryFrom<HashMap<String, String>>

This commit refactors the TryFrom<HashMap<String, String>> implementation for the Task struct in order to handle the previous_task_id field. Previously, if the previous_task_id value was "null", it was treated as None. However, this caused issues when parsing the value as a u64.

Now, the code explicitly checks if the previous_task_id is "null" and sets it to None. Otherwise, it attempts to parse the value as a u64. If parsing fails, an error is logged and the value is set to None.

This change improves the robustness and error handling of the TryFrom implementation, ensuring that the previous_task_id field is correctly handled in all scenarios.

* fix: Surface errors when combining JSON arrays into a single array in get_task_data fails

This change introduces a new `combine_data` function that takes a vector of JSON strings and returns a result containing the combined array and surfaces cases where the parsing failures rather than silently failing.

The `combine_data` function iterates over each JSON string, parses it into a vector of `serde_json::Value`, and then extends the accumulator vector with the parsed values. If parsing fails for any JSON string, an error is logged and returned.

* feat: Update react-query dependency to version 5

Update the react-query dependency in the frontend package.json file to version 5. This update replaces the previous version 3.39.3 with the new version 5, which is provided by the "@tanstack/react-query" package. This update ensures compatibility with the latest features and improvements of react-query.

* Refactor error handling in YouTube API handlers

* Update rust.yml

* Add missing line